moved handling of websocket to utils
