+++ server/src/main.rs
@@ -17,6 +17,8 @@ use watertank_simulation_server::utils::protocol;
 
 use watertank_simulation_server::utils::simulation;
 
+use watertank_simulation_server::utils::server;
+
 use std::{thread, time};
 
 #[tokio::main]
@@ -76,30 +78,7 @@ async fn main() {
 async fn listen_ws(listener: TcpListener) {
     tokio::spawn(async move {
         while let Ok((stream, _)) = listener.accept().await {
-            handle_ws(stream).await;
-        }
-    });
-}
-
-async fn handle_ws(stream: TcpStream) {
-    tokio::spawn(async move {
-        let addr = stream.peer_addr().expect("connected streams should have a peer address");
-        debug!("Peer address: {}", addr);
-
-        let ws_stream = tokio_tungstenite::accept_async(stream)
-            .await
-            .expect("Error during the websocket handshake occurred");
-
-        debug!("New WebSocket connection: {}", addr);
-
-        let (mut write, _) = ws_stream.split();
-        let mut i = 0;
-
-        loop {
-            sleep(Duration::from_millis(100)).await;
-            i += 1;
-            let message = Message::Text(i.to_string());
-            write.send(message).await.unwrap();
+            server::handle_ws(stream).await;
         }
     });
 }
